Refactor: 무한스크롤 팔로잉페이지에서도 사용가능하도록 수정 및 shared로 분리
